Move generator files
