+++ src/pages/lessons/custom/generator/components/RuleCheckbox.tsx
