+++ src/pages/lessons/custom/generator/utilities/rulesCheckboxes.ts
